Change class names
- Change DAO layer interfaces names
- Remove duplicate of DAO class
